refactor: split to introduce tests
